requirements.txt
apache-airflow==2.8.1
requests==2.31.0
pandas==2.1.0
boto3==1.34.0
psycopg2-binary==2.9.9
dbt-postgres==1.7.0
streamlit==1.31.0
plotly==5.18.0
python-dotenv==1.0.0
scipy==1.11.0